Extract frame metadata CSV tokenizer
