+++ client/dive-common/frameMetadata/csvTokenizer.ts
@@ -1,0 +1,91 @@
+type DelimitedTableDelimiter = ',' | '\t';
+
+// Match Python's csv.field_size_limit() default so browser and desktop reject pathological cells
+// consistently.
+const FIELD_SIZE_LIMIT = 131072;
+
+// Keep tokenization node-free because frame metadata parsing runs in both Electron and the browser
+// renderer. The parser is lenient with bare quotes because field logs commonly contain units such
+// as `5"`.
+function tokenizeDelimited(text: string, delimiter: DelimitedTableDelimiter): string[][] {
+  const rows: string[][] = [];
+  let row: string[] = [];
+  let field = '';
+  let inQuotes = false;
+  let atFieldStart = true;
+  let recordHasContent = false;
+  const endField = () => {
+    row.push(field);
+    field = '';
+    atFieldStart = true;
+  };
+  const endRow = () => {
+    endField();
+    rows.push(row);
+    row = [];
+    recordHasContent = false;
+  };
+  const { length } = text;
+  let i = 0;
+  while (i < length) {
+    const ch = text[i];
+    if (inQuotes) {
+      if (ch === '"') {
+        if (text[i + 1] === '"') {
+          field += '"';
+          i += 2;
+        } else {
+          inQuotes = false;
+          i += 1;
+        }
+      } else {
+        field += ch;
+        i += 1;
+      }
+    } else if (ch === '"' && atFieldStart) {
+      inQuotes = true;
+      atFieldStart = false;
+      recordHasContent = true;
+      i += 1;
+    } else if (ch === delimiter) {
+      recordHasContent = true;
+      endField();
+      i += 1;
+    } else if (ch === '\n' || ch === '\r') {
+      if (ch === '\r' && text[i + 1] === '\n') {
+        i += 1;
+      }
+      endRow();
+      i += 1;
+    } else {
+      field += ch;
+      atFieldStart = false;
+      recordHasContent = true;
+      i += 1;
+    }
+  }
+  if (recordHasContent || field.length > 0 || row.length > 0) {
+    endField();
+    rows.push(row);
+  }
+  return rows;
+}
+
+function parseDelimitedRows(text: string, delimiter: DelimitedTableDelimiter): string[][] {
+  if (text.includes('\0')) {
+    throw new Error('line contains NUL');
+  }
+  const rows = tokenizeDelimited(text, delimiter);
+  if (rows.some((row) => row.some((cell) => cell.length > FIELD_SIZE_LIMIT))) {
+    throw new Error('field larger than the size limit');
+  }
+  return rows.map((row) => row.map((cell) => cell.trim()));
+}
+
+export {
+  parseDelimitedRows,
+};
+
+export type {
+  DelimitedTableDelimiter,
+};
